setuptools: add barebones `setup.py` to use custom version scheme
