+++ .gitignore
@@ -17,7 +17,7 @@ __pycache__
 /AppDir/
 /System Volume Information/
 /test_files/
-rare/__version__.py
+rare/_version.py
 # Nuitka build artifacts
 /rare.build
 /rare.dist
