+++ pyproject.toml
@@ -76,7 +76,7 @@ include-package-data = false
 [tool.setuptools.packages.find]
 where = [ "." ]
 include = [ "rare*" ]
-exclude = [ "rare.mkrelease", ".github*", ".tx*", "docs*" ]
+exclude = [ ".github*", ".tx*", "docs*" ]
 namespaces = true
 
 [tool.setuptools.package-data]
@@ -90,9 +90,9 @@ version = { attr = "rare.__version__" }
 # use only annotated tags ( omit `--tags` argument from `git describe`)
 git_describe_command = [ "git", "describe", "--dirty", "--long" ]
 #tag_regex = "(?P<version>^[0-9]+\\.[0-9]+\\.[0-9]+)(?P<suffix>.[0-9]+)?$"
-version_scheme = "rare.mkrelease:version"
+version_scheme = "tools.version:mkversion"
 local_scheme = "no-local-version"
-version_file = "rare/__version__.py"
+version_file = "rare/_version.py"
 fallback_version = "1.10.11.0"
 
 [tool.nuitka]
